Move ParquetLoader to a separate submodule
