+++ packages/evo-objects/src/evo/objects/loader/parquet_loader.py
@@ -1,9 +1,22 @@
+#  Copyright © 2025 Bentley Systems, Incorporated
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#      http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 from __future__ import annotations
 
 from io import BytesIO
 from logging import getLogger
 from typing import cast
 
+import pyarrow as pa
+import pyarrow.parquet as pq
 from pydantic import TypeAdapter
 
 from evo.common import ICache, IFeedback, ITransport
@@ -11,14 +24,8 @@ from evo.common.io import BytesDestination, ChunkedIOManager, Download, HTTPSour
 from evo.common.utils import NoFeedback
 
 from ..exceptions import SchemaValidationError
-from . import ArrowTableFormat, KnownTableFormat
+from ..utils import ArrowTableFormat, KnownTableFormat
 from .types import TableInfo
-
-try:
-    import pyarrow as pa
-    import pyarrow.parquet as pq
-except ImportError:
-    raise ImportError("The 'pyarrow' package is required to use ParquetLoader") from None
 
 try:
     import pandas as pd
